uprec business logic relocation
