+++ db/CMakeLists.txt
@@ -93,6 +93,7 @@ set(src
   dohsql.c
   dohast.c
   machclass.c
+  osqluprec.c
   ${PROJECT_BINARY_DIR}/protobuf/bpfunc.pb-c.c
   ${PROJECT_SOURCE_DIR}/tools/cdb2_dump/cdb2_dump.c
   ${PROJECT_SOURCE_DIR}/tools/cdb2_load/cdb2_load.c
